# [16613] speed-up parquet.write_metadata