--- a/python/pyarrow/parquet/__init__.py
+++ b/python/pyarrow/parquet/__init__.py
@@ -3297,8 +3297,7 @@ def write_metadata(schema, where, metadata_collector=None, **kwargs):
         # ParquetWriter doesn't expose the metadata until it's written. Write
         # it and read it again.
         metadata = read_metadata(where)
-        for m in metadata_collector:
-            metadata.append_row_groups(m)
+        metadata.append_row_groups_list(metadata_collector)
         metadata.write_metadata_file(where)