@AlenkaF 2022-05-26
Thanks @kshitij12345 for contributing!

Current work looks perfect. Interested to see the continuation of it, join of two cython functions into one to accept one or a list of `FileMetaData` objects and also the tests.

Could you change the name of the PR to `ARROW-16613: [Python][Parquet] pyarrow.parquet.write_metadata with metadata_collector appears to be O(n^2)` so it connects to the JIRA ticket you are working on?

And I will ask for the workflows to get started on this PR.